+++ lwesp/src/apps/mqtt/lwesp_mqtt_client.c
@@ -304,7 +304,7 @@ prv_write_fixed_header(lwesp_mqtt_client_p client, mqtt_msg_type_t type, uint8_t
     lwesp_buff_write(&client->tx_buff, &b, 1);  /* Write start of packet parameters */
 
     LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE,
-                 "[MQTT] Writing packet type %s to output buffer\r\n", prv_mqtt_msg_type_to_str(type));
+                 "[LWESP MQTT] Writing packet type %s to output buffer\r\n", prv_mqtt_msg_type_to_str(type));
 
     do {                                        /* Encode length, we must write a len byte even if 0 */
         /*
@@ -386,11 +386,11 @@ prv_write_ack_rec_rel_resp(lwesp_mqtt_client_p client, mqtt_msg_type_t msg_type,
         prv_write_u16(client, pkt_id);          /* Write packet ID */
         prv_send_data(client);                  /* Flush data to output */
         LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE,
-                     "[MQTT] Response %s written to output memory\r\n", prv_mqtt_msg_type_to_str(msg_type));
+                     "[LWESP MQTT] Response %s written to output memory\r\n", prv_mqtt_msg_type_to_str(msg_type));
         return 1;
     } else {
         LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE,
-                     "[MQTT] No memory to write %s packet\r\n", prv_mqtt_msg_type_to_str(msg_type));
+                     "[LWESP MQTT] No memory to write %s packet\r\n", prv_mqtt_msg_type_to_str(msg_type));
     }
     return 0;
 }
@@ -429,7 +429,7 @@ prv_send_data(lwesp_mqtt_client_p client) {
             client->is_sending = 1;             /* Remember active sending flag */
         } else {
             LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE_WARNING,
-                         "[MQTT] Cannot send data with error: %d\r\n", (int)res);
+                         "[LWESP MQTT] Cannot send data with error: %d\r\n", (int)res);
         }
     } else {
         /*
@@ -528,7 +528,7 @@ prv_mqtt_process_incoming_message(lwesp_mqtt_client_p client) {
 
     /* Debug message */
     LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_STATE,
-                 "[MQTT] Processing packet type %s\r\n", prv_mqtt_msg_type_to_str(msg_type));
+                 "[LWESP MQTT] Processing packet type %s\r\n", prv_mqtt_msg_type_to_str(msg_type));
 
     /* Check received packet type */
     switch (msg_type) {
@@ -539,7 +539,7 @@ prv_mqtt_process_incoming_message(lwesp_mqtt_client_p client) {
                     client->conn_state = LWESP_MQTT_CONNECTED;
                 }
                 LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE,
-                             "[MQTT] CONNACK received with result: %d\r\n", (int)err);
+                             "[LWESP MQTT] CONNACK received with result: %d\r\n", (int)err);
 
                 /* Notify user layer */
                 client->evt.type = LWESP_MQTT_EVT_CONNECT;
@@ -548,7 +548,7 @@ prv_mqtt_process_incoming_message(lwesp_mqtt_client_p client) {
             } else {
                 /* Protocol violation here */
                 LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE,
-                             "[MQTT] Protocol violation. CONNACK received when already connected!\r\n");
+                             "[LWESP MQTT] Protocol violation. CONNACK received when already connected!\r\n");
             }
             break;
         }
@@ -574,7 +574,7 @@ prv_mqtt_process_incoming_message(lwesp_mqtt_client_p client) {
             data_len = client->msg_rem_len - (data - client->rx_buff);  /* Calculate length of remaining data */
 
             LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE,
-                         "[MQTT] Publish packet received on topic %.*s; QoS: %d; pkt_id: %d; data_len: %d\r\n",
+                         "[LWESP MQTT] Publish packet received on topic %.*s; QoS: %d; pkt_id: %d; data_len: %d\r\n",
                          (int)topic_len, (const char*)topic, (int)qos, (int)pkt_id, (int)data_len);
 
             /*
@@ -586,7 +586,7 @@ prv_mqtt_process_incoming_message(lwesp_mqtt_client_p client) {
              */
             if (qos > 0) {                      /* We have to reply on QoS > 0 */
                 mqtt_msg_type_t rlwesp_msg_type = qos == 1 ? MQTT_MSG_TYPE_PUBACK : MQTT_MSG_TYPE_PUBREC;
-                LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE, "[MQTT] Sending publish resp: %s on pkt_id: %d\r\n",
+                LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE, "[LWESP MQTT] Sending publish resp: %s on pkt_id: %d\r\n",
                              prv_mqtt_msg_type_to_str(rlwesp_msg_type), (int)pkt_id);
                 prv_write_ack_rec_rel_resp(client, rlwesp_msg_type, pkt_id, qos);
             }
@@ -603,7 +603,7 @@ prv_mqtt_process_incoming_message(lwesp_mqtt_client_p client) {
             break;
         }
         case MQTT_MSG_TYPE_PINGRESP: {          /* Respond to PINGREQ received */
-            LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE, "[MQTT] Ping response received\r\n");
+            LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE, "[LWESP MQTT] Ping response received\r\n");
 
             client->evt.type = LWESP_MQTT_EVT_KEEP_ALIVE;
             client->evt_fn(client, &client->evt);
@@ -657,7 +657,7 @@ prv_mqtt_process_incoming_message(lwesp_mqtt_client_p client) {
                 } else {
                     /* Protocol violation at this point! */
                     LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE,
-                                 "[MQTT] Protocol violation. Received ACK without sent packet\r\n");
+                                 "[LWESP MQTT] Protocol violation. Received ACK without sent packet\r\n");
                 }
             }
             break;
@@ -690,7 +690,7 @@ prv_mqtt_parse_incoming(lwesp_mqtt_client_p client, lwesp_pbuf_p pbuf) {
             switch (client->parser_state) {     /* Check parser state */
                 case MQTT_PARSER_STATE_INIT: {  /* We are waiting for start byte and packet type */
                     LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_STATE,
-                                 "[MQTT] Parser init state, received first byte of packet 0x%02X\r\n", (unsigned)ch);
+                                 "[LWESP MQTT] Parser init state, received first byte of packet 0x%02X\r\n", (unsigned)ch);
 
                     /* Save other info about message */
                     client->msg_hdr_byte = ch;  /* Save first entry */
@@ -708,7 +708,7 @@ prv_mqtt_parse_incoming(lwesp_mqtt_client_p client, lwesp_pbuf_p pbuf) {
 
                     if (!(ch & 0x80)) {         /* Is this last entry? */
                         LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_STATE,
-                                     "[MQTT] Remaining length received: %d bytes\r\n", (int)client->msg_rem_len);
+                                     "[LWESP MQTT] Remaining length received: %d bytes\r\n", (int)client->msg_rem_len);
 
                         if (client->msg_rem_len > 0) {
                             /*
@@ -753,12 +753,12 @@ prv_mqtt_parse_incoming(lwesp_mqtt_client_p client, lwesp_pbuf_p pbuf) {
                     if (client->msg_curr_pos == client->msg_rem_len) {
                         if (client->msg_curr_pos <= client->rx_buff_len) {  /* Check if it was possible to write all data to rx buffer */
                             LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_STATE,
-                                         "[MQTT] Packet parsed and ready for processing\r\n");
+                                         "[LWESP MQTT] Packet parsed and ready for processing\r\n");
 
                             prv_mqtt_process_incoming_message(client);  /* Process incoming packet */
                         } else {
                             LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE_WARNING,
-                                         "[MQTT] Packet too big for rx buffer. Packet discarded\r\n");
+                                         "[LWESP MQTT] Packet too big for rx buffer. Packet discarded\r\n");
                         }
                         client->parser_state = MQTT_PARSER_STATE_INIT;  /* Go to initial state and listen for next received packet */
                     }
@@ -892,7 +892,7 @@ prv_mqtt_data_sent_cb(lwesp_mqtt_client_p client, size_t sent_len, uint8_t succe
     if (!successful) {
         prv_mqtt_close(client);
         LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE_WARNING,
-                     "[MQTT] Failed to send %d bytes. Manually closing down..\r\n", (int)sent_len);
+                     "[LWESP MQTT] Failed to send %d bytes. Manually closing down..\r\n", (int)sent_len);
         return 0;
     }
     lwesp_buff_skip(&client->tx_buff, sent_len);/* Skip buffer for actual sent data */
@@ -948,9 +948,9 @@ prv_mqtt_poll_cb(lwesp_mqtt_client_p client) {
             prv_send_data(client);              /* Force send data */
             client->poll_time = 0;              /* Reset polling time */
 
-            LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE, "[MQTT] Sending PINGREQ packet\r\n");
+            LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE, "[LWESP MQTT] Sending PINGREQ packet\r\n");
         } else {
-            LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE_WARNING, "[MQTT] No memory to send PINGREQ packet\r\n");
+            LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE_WARNING, "[LWESP MQTT] No memory to send PINGREQ packet\r\n");
         }
     }
 
@@ -1263,13 +1263,13 @@ lwesp_mqtt_client_publish(lwesp_mqtt_client_p client, const char* topic, const v
             prv_request_set_pending(client, request);   /* Set request as pending waiting for server reply */
             prv_send_data(client);              /* Try to send data */
             LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE,
-                         "[MQTT] Pkt publish start. QoS: %d, pkt_id: %d\r\n", (int)qos_u8, (int)pkt_id);
+                         "[LWESP MQTT] Pkt publish start. QoS: %d, pkt_id: %d\r\n", (int)qos_u8, (int)pkt_id);
         } else {
-            LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE, "[MQTT] No free request available to publish message\r\n");
+            LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE, "[LWESP MQTT] No free request available to publish message\r\n");
             res = lwespERRMEM;
         }
     } else {
-        LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE, "[MQTT] Not enough memory to publish message\r\n");
+        LWESP_DEBUGF(LWESP_CFG_DBG_MQTT_TRACE, "[LWESP MQTT] Not enough memory to publish message\r\n");
         res = lwespERRMEM;
     }
     lwesp_core_unlock();