+++ lwesp/src/lwesp/lwesp_threads.c
@@ -53,7 +53,7 @@ lwesp_thread_produce(void* const arg) {
     uint32_t time;
 
     LWESP_DEBUGF(LWESP_CFG_DBG_THREAD | LWESP_DBG_TYPE_TRACE,
-                 "[THREAD] Producer thread started\r\n");
+                 "[LWESP THREAD] Producer thread started\r\n");
 
     /*
      * Thread parameter is semaphore handle object
@@ -83,7 +83,7 @@ lwesp_thread_produce(void* const arg) {
          */
         if (!e->status.f.dev_present) {
             LWESP_DEBUGF(LWESP_CFG_DBG_THREAD | LWESP_DBG_TYPE_TRACE | LWESP_DBG_LVL_WARNING,
-                         "[THREAD] Device is not present\r\n");
+                         "[LWESP THREAD] Device is not present\r\n");
             res = lwespERRNODEVICE;
         }
 
@@ -127,10 +127,10 @@ lwesp_thread_produce(void* const arg) {
 
             LWESP_DEBUGW(LWESP_CFG_DBG_THREAD | LWESP_DBG_TYPE_TRACE | LWESP_DBG_LVL_SEVERE,
                          res == lwespTIMEOUT,
-                         "[THREAD] Timeout in produce thread waiting for command to finish in process thread\r\n");
+                         "[LWESP THREAD] Timeout in produce thread waiting for command to finish in process thread\r\n");
             LWESP_DEBUGW(LWESP_CFG_DBG_THREAD | LWESP_DBG_TYPE_TRACE | LWESP_DBG_LVL_SEVERE,
                          res != lwespOK && res != lwespTIMEOUT,
-                         "[THREAD] Could not start execution for command %d\r\n", (int)msg->cmd);
+                         "[LWESP THREAD] Could not start execution for command %d\r\n", (int)msg->cmd);
 
             /*
              * Manually release semaphore in all cases:
@@ -199,7 +199,7 @@ lwesp_thread_process(void* const arg) {
     uint32_t time;
 
     LWESP_DEBUGF(LWESP_CFG_DBG_THREAD | LWESP_DBG_TYPE_TRACE,
-                 "[THREAD] Process thread started\r\n");
+                 "[LWESP THREAD] Process thread started\r\n");
 
     /*
      * Thread parameter is semaphore handle object